quad layers v1: promote one opaque overlay quad to its own stream
One quad layer (largest on-screen solid angle, 25% stickiness) is taken
out of the server-side squash, resampled into a dedicated 1024^2 YCbCr
image and encoded as stream 3; the headset decodes it and submits a
real XrCompositionLayerQuad with the server pose, so panels are
runtime-timewarped at display rate and skip foveation/eye-scaling
entirely. Promoting also restores the single-projection fast path and
stops off-axis quads widening the encoded FOV.

Wire: per-stream geometry (quad_width/height + stream_size accessor),
quad placement as an optional quad_info on shard 0 of stream 3.
Blended quads are not promoted in v1 (stream carries no alpha);
quad-layers.allow-blended overrides. Falls back byte-identically to
baking in every uncertain case. Headset toggle 'Sharp overlay layers'.
Also fixes a latent shard_accumulator frame_diff truncation (gap of
exactly 256 frames misfiled shards). Wire tests in tests/.

+++ client/decoder/shard_accumulator.cpp
@@ -113,7 +113,11 @@ void shard_accumulator::push_shard(video_stream_data_shard && shard)
 {
 	assert(current.frame_index() + 1 == next.frame_index());
 
-	uint8_t frame_diff = shard.frame_idx - current.frame_index();
+	// Not truncated to 8 bits: a stream that is silent for a while, as the quad
+	// layer stream is whenever no layer is promoted, comes back with a gap of any
+	// size, and a gap that happened to be a multiple of 256 would look like no gap
+	// at all and file the new frame's shards under the old frame index.
+	uint64_t frame_diff = shard.frame_idx - current.frame_index();
 	if (shard.frame_idx < current.frame_index())
 	{
 		// frame is in the past, drop it
